feat(ui): add NotificationProvider
