+++ apps/site/components/Common/CodeBox.tsx
@@ -3,11 +3,12 @@
 import { CodeBracketIcon } from '@heroicons/react/24/outline';
 import BaseCodeBox from '@node-core/ui-components/Common/BaseCodeBox';
 import styles from '@node-core/ui-components/Common/BaseCodeBox/index.module.css';
+import { useNotification } from '@node-core/ui-components/Providers/NotificationProvider';
 import { useTranslations } from 'next-intl';
 import type { FC, PropsWithChildren } from 'react';
 
 import Link from '#site/components/Link';
-import { useCopyToClipboard, useNotification } from '#site/hooks';
+import { useCopyToClipboard } from '#site/hooks';
 
 type CodeBoxProps = {
   language: string;
