+++ apps/site/hooks/react-client/index.ts
@@ -1,6 +1,5 @@
 export { default as useCopyToClipboard } from './useCopyToClipboard';
 export { default as useDetectOS } from './useDetectOS';
 export { default as useMediaQuery } from './useMediaQuery';
-export { default as useNotification } from './useNotification';
 export { default as useClientContext } from './useClientContext';
 export { default as useNavigationState } from './useNavigationState';
