--- apps/site/hooks/react-client/useNotification.ts
@@ -1,9 +1,0 @@
-'use client';
-
-import { useContext } from 'react';
-
-import { NotificationDispatch } from '#site/providers/notificationProvider';
-
-const useNotification = () => useContext(NotificationDispatch);
-
-export default useNotification;
