+++ apps/site/hooks/react-server/index.ts
@@ -1,5 +1,4 @@
 export { default as useCopyToClipboard } from './useCopyToClipboard';
 export { default as useDetectOS } from './useDetectOS';
 export { default as useMediaQuery } from './useMediaQuery';
-export { default as useNotification } from './useNotification';
 export { default as useClientContext } from './useClientContext';
