--- apps/site/hooks/react-server/useNotification.ts
@@ -1,5 +1,0 @@
-const useNotification = () => {
-  throw new Error('Attempted to call useNotification from RSC');
-};
-
-export default useNotification;
