+++ apps/site/layouts/Base.tsx
@@ -1,9 +1,9 @@
 'use client';
 
+import { NotificationProvider } from '@node-core/ui-components/Providers/NotificationProvider';
 import type { FC, PropsWithChildren } from 'react';
 
 import { NavigationStateProvider } from '#site/providers/navigationStateProvider';
-import { NotificationProvider } from '#site/providers/notificationProvider';
 
 import styles from './layouts.module.css';
