+++ apps/site/package.json
@@ -44,7 +44,6 @@
     "@oramacloud/client": "^2.1.4",
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-tabs": "^1.1.12",
-    "@radix-ui/react-toast": "^1.2.14",
     "@radix-ui/react-tooltip": "^1.2.7",
     "@tailwindcss/postcss": "~4.1.11",
     "@types/node": "catalog:",
